Avoid expensive Project.setPaths calls before each test
Project.setPaths is expensive because it creates and destrosy file
watchers and, even worse, git repositories. We need to reset
atom.project after each test. But the vast majority of the time, we
reset the project to use same paths as before.

Now, we pass the desired project paths to AtomEnvironment.reset
after each test so that we can avoid spuriously clearing and
re-populating the paths.